spring show working
